Исправлены ошибки.

+++ src/main/java/jm/task/core/jdbc/dao/UserDaoHibernateImpl.java
@@ -7,11 +7,12 @@ import org.hibernate.SessionFactory;
 import org.hibernate.Transaction;
 
 
+import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.List;
 
 
-
+@Transactional
 public class UserDaoHibernateImpl implements UserDao {
 
     private final SessionFactory sessionFactory = Util.getSessionFactory();
@@ -23,44 +24,38 @@ public class UserDaoHibernateImpl implements UserDao {
 
     @Override
     public void createUsersTable() {
-        Transaction transaction = null;
+        Transaction transaction;
         try (Session session = sessionFactory.openSession()) {
             transaction = session.beginTransaction();
             session.createSQLQuery("CREATE TABLE IF NOT EXISTS Users" +
-                    " (id mediumint not null auto_increment, name VARCHAR(50), " +
-                    "lastname VARCHAR(50), " +
+                    " (id bigint not null auto_increment, name text, " +
+                    "lastname text, " +
                     "age tinyint, " +
                     "PRIMARY KEY (id))").executeUpdate();
             transaction.commit();
             System.out.println("Таблица создана");
         } catch (Exception e) {
             e.printStackTrace();
-            if (transaction != null) {
-                transaction.rollback();
-            }
         }
     }
 
     @Override
     public void dropUsersTable() {
-        Transaction transaction = null;
+        Transaction transaction;
         try (Session session = sessionFactory.openSession()) {
-            transaction = null;
             transaction = session.beginTransaction();
             session.createSQLQuery("Drop table if exists Users").executeUpdate();
             transaction.commit();
             System.out.println("Таблица удалена");
         } catch (Exception e) {
             e.printStackTrace();
-            if (transaction != null) {
-                transaction.rollback();
-            }
+
         }
     }
 
     @Override
     public void saveUser(String name, String lastName, byte age) {
-        Transaction transaction = null;
+        Transaction transaction;
         try (Session session = sessionFactory.openSession()) {
             transaction = session.beginTransaction();
             session.save(new User(name, lastName, age));
@@ -68,53 +63,44 @@ public class UserDaoHibernateImpl implements UserDao {
             System.out.println("User с именем – " + name + " добавлен в базу данных");
         } catch (Exception e) {
             e.printStackTrace();
-            if (transaction != null) {
-                transaction.rollback();
-            }
         }
     }
 
     @Override
     public void removeUserById(long id) {
-        Transaction transaction = null;
+        Transaction transaction;
         try (Session session = sessionFactory.openSession()) {
             transaction = session.beginTransaction();
-            session.delete(session.get(User.class, id));
+            if (session.get(User.class, id) != null) session.delete(session.get(User.class, id));
             transaction.commit();
             System.out.println("User удален");
         } catch (Exception e) {
             e.printStackTrace();
-            if (transaction != null) {
-                transaction.rollback();
-            }
         }
     }
 
     @Override
     public List<User> getAllUsers() {
         List <User> list = new ArrayList<>();
-        Transaction transaction = null;
+        Transaction transaction;
         try (Session session = sessionFactory.openSession()) {
             transaction = session.beginTransaction();
             list = session.createCriteria(User.class).list();
             transaction.commit();
         } catch (Exception e) {
             e.printStackTrace();
-            if (transaction != null) {
-                transaction.rollback();
-            }
         }
         return list;
     }
 
     @Override
     public void cleanUsersTable() {
-        Transaction transaction = null;
+        Transaction transaction;
         try (Session session = sessionFactory.openSession()) {
             transaction = session.beginTransaction();
             final List<User> instances = session.createCriteria(User.class).list();
 
-            for (Object o : instances) {
+            for (User o : instances) {
                 session.delete(o);
             }
 
@@ -122,9 +108,6 @@ public class UserDaoHibernateImpl implements UserDao {
             System.out.println("Таблица очищена");
         } catch (Exception e) {
             e.printStackTrace();
-            if (transaction != null) {
-                transaction.rollback();
-            }
         }
     }
 }
